requirements.txt
requests==2.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
numpy==1.26.4
pyTelegramBotAPI==4.21.0
